+++ apis/python/src/tiledbsoma/_tdb_handles.py
@@ -212,20 +212,6 @@ class Wrapper(Generic[_RawHdl_co], metaclass=abc.ABCMeta):
         """Opens and returns a TileDB object specific to this type."""
         raise NotImplementedError()
 
-    def reopen(self, mode: options.OpenMode, timestamp: OpenTimestamp | None) -> Wrapper[_RawHdl_co]:
-        """Returns a new copy of the wrapper handle in the requested mode and at the requested timestamp."""
-        if mode not in ("r", "w"):
-            raise ValueError(f"Invalid mode '{mode}' passed. " "Valid modes are 'r' and 'w'.")
-        ts = self.context._open_timestamp_ms(timestamp)
-        self.metadata._write()
-        clib_handle = self._handle.reopen(clib.OpenMode.read if mode == "r" else clib.OpenMode.write, (0, ts))
-        return self.__class__.open(
-            uri=clib_handle.uri,
-            mode=mode,
-            context=self.context,
-            timestamp=ts,
-        )
-
     # Covariant types should normally not be in parameters, but this is for
     # internal use only so it's OK.
     def _do_initial_reads(self, reader: _RawHdl_co) -> None:  # type: ignore[misc]